+++ src/main/java/io/nats/client/impl/NatsConnectionWriter.java
@@ -23,16 +23,17 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.ReentrantLock;
 
 class NatsConnectionWriter implements Runnable {
 
     private final NatsConnection connection;
 
-    private Thread thread;
-    private CompletableFuture<Boolean> stopped;
+    private Future<Boolean> stopped;
     private Future<DataPort> dataPortFuture;
     private final AtomicBoolean running;
     private final AtomicBoolean reconnectMode;
+    private final ReentrantLock startStopLock;
 
     private byte[] sendBuffer;
 
@@ -44,8 +45,9 @@ class NatsConnectionWriter implements Runnable {
 
         this.running = new AtomicBoolean(false);
         this.reconnectMode = new AtomicBoolean(false);
+        this.startStopLock = new ReentrantLock();
         this.stopped = new CompletableFuture<>();
-        this.stopped.complete(Boolean.TRUE); // we are stopped on creation
+        ((CompletableFuture<Boolean>)this.stopped).complete(Boolean.TRUE); // we are stopped on creation
 
         this.sendBuffer = new byte[connection.getOptions().getBufferSize()];
 
@@ -57,32 +59,41 @@ class NatsConnectionWriter implements Runnable {
     // Use the Future from stop() to determine if it is ok to call this.
     // This method resets that future so mistiming can result in badness.
     void start(Future<DataPort> dataPortFuture) {
-        this.dataPortFuture = dataPortFuture;
-        this.running.set(true);
-        this.stopped = new CompletableFuture<>(); // New future
-
-        String name = (this.connection.getOptions().getConnectionName() != null) ? this.connection.getOptions().getConnectionName() : "Nats Connection";
-        this.thread = new Thread(this, name + " Writer");
-        this.thread.start();
+        this.startStopLock.lock();
+        try {
+            this.dataPortFuture = dataPortFuture;
+            this.running.set(true);
+            this.outgoing.resume();
+            this.reconnectOutgoing.resume();
+            this.stopped = connection.getExecutor().submit(this, Boolean.TRUE);
+        } finally {
+            this.startStopLock.unlock();
+        }
     }
 
     // May be called several times on an error.
     // Returns a future that is completed when the thread completes, not when this
     // method does.
     Future<Boolean> stop() {
-        this.running.set(false);
-        this.outgoing.pause();
-        this.reconnectOutgoing.pause();
-
-        // Clear old ping/pong requests
-        byte[] pingRequest = NatsConnection.OP_PING.getBytes(StandardCharsets.UTF_8);
-        byte[] pongRequest = NatsConnection.OP_PONG.getBytes(StandardCharsets.UTF_8);
-        this.outgoing.filter((msg) -> {
-            return Arrays.equals(pingRequest, msg.getProtocolBytes()) || Arrays.equals(pongRequest, msg.getProtocolBytes());
-        });
+        this.startStopLock.lock();
+        try {
+            this.running.set(false);
+            this.outgoing.pause();
+            this.reconnectOutgoing.pause();
+    
+            // Clear old ping/pong requests
+            byte[] pingRequest = NatsConnection.OP_PING.getBytes(StandardCharsets.UTF_8);
+            byte[] pongRequest = NatsConnection.OP_PONG.getBytes(StandardCharsets.UTF_8);
+            this.outgoing.filter((msg) -> {
+                return Arrays.equals(pingRequest, msg.getProtocolBytes()) || Arrays.equals(pongRequest, msg.getProtocolBytes());
+            });
+        } finally {
+            this.startStopLock.unlock();
+        }
         return this.stopped;
     }
 
+    @Override
     public void run() {
         Duration waitForMessage = Duration.ofMinutes(2); // This can be long since no one is sending
         Duration reconnectWait = Duration.ofMillis(1); // This can be long since no one is sending
@@ -91,8 +102,6 @@ class NatsConnectionWriter implements Runnable {
         try {
             DataPort dataPort = this.dataPortFuture.get(); // Will wait for the future to complete
             NatsStatistics stats = this.connection.getNatsStatistics();
-            this.outgoing.resume();
-            this.reconnectOutgoing.resume();
 
             while (this.running.get()) {
                 int sendPosition = 0;
@@ -157,8 +166,6 @@ class NatsConnectionWriter implements Runnable {
             // Exit
         } finally {
             this.running.set(false);
-            this.stopped.complete(Boolean.TRUE);
-            this.thread = null;
         }
     }