+++ src/main/java/io/nats/client/impl/NatsConsumer.java
@@ -201,7 +201,7 @@ abstract class NatsConsumer implements Consumer {
 
         // Wait for the timeout or the pending count to go to 0, skipped if conn is
         // draining
-        Thread t = new Thread(() -> {
+        connection.getExecutor().submit(() -> {
             try {
                 Instant now = Instant.now();
 
@@ -223,8 +223,6 @@ abstract class NatsConsumer implements Consumer {
                 tracker.complete(this.isDrained());
             }
        });
-       t.setName("Consumer Drain");
-       t.start();
 
        return getDrainingFuture();
    }