+++ src/test/java/io/nats/client/OptionsTests.java
@@ -27,6 +27,11 @@ import java.time.Duration;
 import java.util.Base64;
 import java.util.Collection;
 import java.util.Properties;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.SSLContext;
 
@@ -426,6 +431,20 @@ public class OptionsTests {
     }
 
     @Test
+    public void testServersWithCommas() {
+        String url1 = "nats://localhost:8080";
+        String url2 = "nats://localhost:8081";
+        String serverURLs = url1 + "," + url2;
+        Options o = new Options.Builder().server(serverURLs).build();
+
+        Collection<URI> servers = o.getServers();
+        URI[] serverArray = servers.toArray(new URI[0]);
+        assertEquals(2, serverArray.length);
+        assertEquals("property server", url1, serverArray[0].toString());
+        assertEquals("property server", url2, serverArray[1].toString());
+    }
+
+    @Test
     public void testEmptyStringInServers() {
         String url1 = "nats://localhost:8080";
         String url2 = "";
@@ -474,6 +493,34 @@ public class OptionsTests {
         String[] serverUrls = {url1, url2};
         new Options.Builder().servers(serverUrls).build();
         assertFalse(true);
+    }
+    
+    @Test
+    public void testSetExectuor() {
+        ExecutorService exec = Executors.newCachedThreadPool();
+        Options options = new Options.Builder().executor(exec).build();
+        assertEquals(exec, options.getExecutor());
+    }
+    
+    @Test
+    public void testDefaultExecutor() throws Exception {
+        Options options = new Options.Builder().connectionName("test").build();
+        Future<String> future = options.getExecutor().submit(new Callable<String>(){
+            public String call() {
+                return Thread.currentThread().getName();
+            }
+        });
+        String name = future.get(5, TimeUnit.SECONDS);
+        assertTrue(name.startsWith("test"));
+
+        options = new Options.Builder().build();
+        future = options.getExecutor().submit(new Callable<String>(){
+            public String call() {
+                return Thread.currentThread().getName();
+            }
+        });
+        name = future.get(5, TimeUnit.SECONDS);
+        assertTrue(name.startsWith(Options.DEFAULT_THREAD_NAME_PREFIX));
     }
 
     @Test